Normal Distributions
The normal distribution is a suitable model for many types of data. Assuming normality without cause can generate problems in modelling. Generally it is useful to test the normality of an unknown dataset.
Study the given data comprising two variables x1 and x2. Test each variable separately for normality with the following statistical tests and 5% significance level:
Nonparametric probability density estimation can be implemented based on the nearest-neighbour approach. In this case, the density estimate p^(x)=kNV(x)
 where k is a fixed number of neighbours, N is the number of samples and V the (local) length/area/volume that contains the specified number of samples.

Implement a nearest-neighbour estimator in Matlab