Update and rename hwsixtwo.m to Normal Distribution
Nonparametric probability density estimation can be implemented based on the nearest-neighbour approach. In this case, the density estimate p^(x)=kNV(x)
 where k is a fixed number of neighbours, N is the number of samples and V the (local) length/area/volume that contains the specified number of samples.

Implement a nearest-neighbour estimator in Matlab/Python.
Use the provided data for experimentation with the estimator. For having a distribution to compare with, use the normal distribution and estimate its parameters from the data. The best value for  k